fix(sync): align sync v2 store layering
